LPS-194619 Use a unique key
kaleoTaskAssignmentId is attached to the workflow definition, not the specific instance of that workflow in action.
kaleoInstanceTokenId is the same for each workflow instance but it's reused from one task to another.
currentKaleoNodeId specifies the exact task node, making the combined key unique.

Note that _kaleoTaskInstanceToken from executionContext is null at this point so it cannot be used as key.